requirements.txt
ipykernel 
langchain
langchain_community 
langchain_groq
langgraph
pydantic
langchain_core
typing_extensions
arxiv
wikipedia
PyPDF2
playwright
reportlab
fpdf
graphviz
tensorflow
bs4
requests
streamlit
google-generativeai
pdf2image